refactor(util): split RFDBServerBackend god-file into layered modules (REG-490)
RFDBServerBackend.ts had grown to 1035 lines — 2x over the repo's hard
500-line limit (and 66% over the 700-line CRITICAL threshold cited in the
issue). Split the single class into a behavior-preserving layered chain plus
a pure codec module:

- rfdbTypes.ts (80) — shared public interfaces (re-exported from the backend
  for back-compat).
- rfdbCodec.ts (277) — pure, stateless helpers extracted verbatim: socket-path
  resolution, PID polling, node/edge wire conversion, input validation, record
  parsing and Datalog binding normalization.
- RFDBConnectionBase.ts (246) — connection lifecycle + shared protected state.
- RFDBDataBackend.ts (345) — node/edge CRUD, traversal and batch ops.
- RFDBServerBackend.ts (230) — public class: statistics, Datalog, Cypher; the
  full method surface remains on RFDBServerBackend via the inheritance chain
  (RFDBConnectionBase -> RFDBDataBackend -> RFDBServerBackend), so all ~55
  importers and the default export are unchanged.

Behavior is byte-for-byte preserved: the subtle divergence where addEdges()
stashes _origSrc/_origDst into v2 metadata but batchEdge() never does is kept
exactly, and all error messages stay verbatim.

Adds test/unit/rfdbCodec.test.js (30 server-free characterization tests) that
lock the extracted logic.

Gate: pnpm build OK, ./scripts/test.sh 771/771 pass (was 741 + 30 new),
pnpm typecheck OK, pnpm lint OK. End-to-end: the gated suite's server-based
tests (e.g. DeferredIndexing) drive the refactored backend against a live
rfdb-server and pass.
